refactor: extract bash_safety.rs from approval.rs (#116 P0-1)
approval.rs had 3 responsibilities: mode logic, bash safety classifier,
and settings persistence. The safety classifier (SAFE_PREFIXES,
DANGEROUS_PATTERNS, is_command_safe, pipeline parsing, whitelist
extraction) is a self-contained mini bash parser with zero coupling
to ApprovalMode.

Extracted to bash_safety.rs:
- All safety classification logic (~450 lines)
- 23 tests (12 existing + 11 moved from approval.rs)
- extract_whitelist_pattern (re-exported from approval for inference.rs)

approval.rs: 1,127 → 464 lines
bash_safety.rs: 672 lines (self-contained)

All tests pass, clippy clean.
